+++ feature/expo/src/main/java/com/school_of_company/expo/view/component/DetailsPieChartItem.kt
@@ -1,0 +1,60 @@
+package com.school_of_company.expo.view.component
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+import com.school_of_company.design_system.theme.ExpoTypography
+import com.school_of_company.design_system.theme.color.ExpoColor
+
+@Composable
+internal fun DetailsPieChartItem(
+    label: String,
+    value: Int,
+    percent: Int,
+    color: Color
+) {
+    Row(
+        modifier = Modifier.fillMaxWidth(),
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        Box(
+            modifier = Modifier
+                .background(
+                    color = color,
+                    shape = CircleShape
+                )
+                .size(16.dp)
+        )
+
+        Column(modifier = Modifier.fillMaxWidth()) {
+            Text(
+                modifier = Modifier.padding(start = 15.dp),
+                text = label,
+                style = ExpoTypography.captionBold1,
+                color = Color.Black
+            )
+
+            Spacer(modifier = Modifier.height(8.dp))
+
+            Text(
+                modifier = Modifier.padding(start = 15.dp),
+                text = "${value}명 ($percent%)",
+                style = ExpoTypography.captionRegular2,
+                color = ExpoColor.gray400
+            )
+        }
+    }
+}
